Fix recycler view progress indicator
The recycler view didn't render new items until a scroll event occurred.
This is fixed by calling the notifyItemRangeInserted method to render
new item views in a granular manner.

A fatal exception was thrown by the recycler view after its adapter was
told to remove the progress indicator view. This is fixed by calling the
notifyItemRemoved method with the index of the removed view.

The adapter wont show the progress indicator view when the adapter is
empty as this makes EndlessRecyclerViewScrollListener behave
unexpectedly.